Memora Note Making App for Lab Exam 4 Update some new feature

+++ app/src/main/java/com/example/labexam4/NotesAdapter.kt
@@ -1,5 +1,6 @@
 package com.example.labexam4
 
+import android.app.AlertDialog
 import android.content.Context
 import android.content.Intent
 import android.view.LayoutInflater
@@ -41,11 +42,20 @@ class NotesAdapter(private var notes: List<Note>, context: Context) : RecyclerVi
         }
 
         holder.deleteButton.setOnClickListener {
-            db.deleteNote(note.id)
-            refreshData(db.getAllNotes())
-            Toast.makeText(holder.itemView.context,"Note Deleted",Toast.LENGTH_SHORT).show()
+            val context = holder.itemView.context
+            AlertDialog.Builder(context).apply {
+                setTitle("Delete Note")
+                setMessage("Are you sure you want to delete this note?")
+                setPositiveButton("OK") { dialog, which ->
+                    db.deleteNote(note.id)
+                    refreshData(db.getAllNotes())
+                    Toast.makeText(context, "Note Deleted", Toast.LENGTH_SHORT).show()
+                }
+                setNegativeButton("Cancel") { dialog, which ->
+                    dialog.dismiss()
+                }
+            }.create().show()
         }
-
     }
 
     fun refreshData(newNotes: List<Note>) {
